feature: tab view and navigation

+++ ManageDose/Components/PieChart.swift
@@ -27,9 +27,9 @@ struct PieChart: View {
     let basic: Double
     let savings: Double
     let wants: Double
-    let budget: Int
+    let budget: Double
     
-    init(basic: Double, savings: Double, wants: Double, budget: Int) {
+    init(basic: Double, savings: Double, wants: Double, budget: Double) {
         self.basic = basic
         self.savings = savings
         self.wants = wants
